visualization_experiments/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
matplotlib>=3.5.0
numpy>=1.21.0
pyyaml>=6.0
scipy>=1.7.0
scikit-learn>=1.0.0
pandas>=1.3.0
tqdm>=4.62.0
